python_service/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart>=0.0.9

# Freelancer SDK
freelancersdk>=0.1.20

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
aiohttp>=3.9.0

# LLM
openai>=1.0.0

# Utilities
python-dateutil==2.8.2
PyYAML==6.0.1
